text: move bbulk and gltex to the render/ directory
