+++ src/kmscon_main.c
@@ -35,12 +35,12 @@
 #include "input/input.h"
 #include "kmscon_conf.h"
 #include "kmscon_seat.h"
+#include "render/text.h"
 #include "shl/dlist.h"
 #include "shl/eloop.h"
 #include "shl/log.h"
 #include "shl/misc.h"
 #include "shl/module.h"
-#include "text.h"
 #include "uterm_monitor.h"
 #include "uterm_vt.h"
 #include "video/video.h"
