+++ src/render/bbulk.c
