+++ src/render/gltex_atlas.frag
