+++ src/render/gltex_atlas.vert
